requirements.txt
python-dotenv
rich
instaloader
pyfiglet

